requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.10.0    
cachetools==5.0.0
certifi==2021.10.8        
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cv2-tools==2.4.0
cycler==0.11.0
Cython==0.29.28
decorator==5.1.1
edit-distance==1.0.4
editdistance-s==1.0.0
executing==0.8.3
Flask==2.1.1
flatbuffers==2.0
fonttools==4.31.2
fpdf==1.7.2
future==0.18.2
gast==0.5.3
gax-cloud-vision-v1==0.8.0
google-api-core==2.7.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-cloud-vision==2.7.2
google-gax==0.12.5
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpc-cloud-vision-v1==0.8.0
grpc-google-cloud-vision-v1==0.14.0
grpcio==1.44.0
grpcio-status==1.44.0
h5py==3.6.0
httplib2==0.20.4
idna==3.3
image-enhancement==0.2.0
ImageFilter==0.1.0
ImageHash==4.2.1
imageio==2.16.1
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
networkx==2.7.1
numpy==1.22.3
oauth2client==3.0.0
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
PIL-Tools==1.1.0
Pillow==9.1.0
ply==3.8
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
python-constraint==1.4.0
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
PyWavelets==1.3.0
pywhatkit==5.3
reload==0.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
soupsieve==2.3.2
stack-data==0.2.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.3.25
traitlets==5.1.1
types-aiofiles==0.8.6
types-all==1.0.0
types-annoy==1.17.5
types-atomicwrites==1.4.3
types-backports==0.1.3
types-backports-abc==0.5.2
types-bleach==4.1.5
types-boto==2.49.11
types-cachetools==5.0.0
types-certifi==2021.10.8.1
types-characteristic==14.3.3
types-chardet==4.0.3
types-click==7.1.8
types-click-spinner==0.1.10
types-colorama==0.4.11
types-contextvars==2.4.4
types-croniter==1.0.8
types-cryptography==3.3.19
types-dataclasses==0.6.4
types-dateparser==1.1.0
types-DateTimeRange==1.2.4
types-decorator==5.1.4
types-Deprecated==1.2.5
types-docopt==0.6.6
types-docutils==0.18.0
types-emoji==1.2.7
types-enum34==1.1.8
types-fb303==1.0.0
types-filelock==3.2.5
types-first==2.0.2
types-Flask==1.1.6
types-freezegun==1.1.8
types-frozendict==2.0.7
types-futures==3.3.8
types-geoip2==3.0.0
types-ipaddress==1.0.8
types-itsdangerous==1.1.6
types-JACK-Client==0.5.5
types-Jinja2==2.11.9
types-kazoo==0.1.3
types-Markdown==3.3.12
types-MarkupSafe==1.1.10
types-maxminddb==1.5.0
types-mock==4.0.13
types-mypy-extensions==0.4.14
types-nmap==0.1.6
types-openssl-python==0.1.3
types-orjson==3.6.2
types-paramiko==2.8.17
types-pathlib2==2.3.0
types-Pillow==9.0.11
types-pkg-resources==0.1.3
types-polib==1.1.11
types-protobuf==3.19.15
types-pyaudio==0.2.11
types-pycurl==7.44.7
types-pyfarmhash==0.2.3
types-PyJWT==1.7.1
types-pymssql==2.1.0
types-PyMySQL==1.0.15
types-pyRFC3339==1.1.1
types-pysftp==0.2.12
types-python-dateutil==2.8.10
types-python-gflags==3.1.4
types-python-slugify==5.0.3
types-pytz==2021.3.6
types-pyvmomi==7.0.7
types-PyYAML==6.0.5
types-redis==4.1.19
types-requests==2.27.16
types-retry==0.9.5
types-Routes==2.5.0
types-scribe==2.0.0
types-simplejson==3.17.4
types-singledispatch==3.7.3
types-six==1.16.12
types-tabulate==0.8.6
types-termcolor==1.1.3
types-toml==0.10.4
types-tornado==5.1.1
types-typed-ast==1.5.2
types-tzlocal==4.1.0
types-ujson==4.2.1
types-urllib3==1.26.11
types-waitress==2.0.6
types-Werkzeug==1.0.9
types-xxhash==2.0.3
typing_extensions==4.1.1
urllib3==1.26.9
watchdog==2.1.7
wcwidth==0.2.5
Werkzeug==2.1.1
wikipedia==1.4.0
wrapt==1.14.0
gunicorn==19.7.1
